Begin compiler
Start of the compiler. The compiler does not directly produce the `.ksm`
binary, but it gives the kRISC instructions to be placed in a binary.
While it is currently far from done, it has output its first program
(with help by manually* assmebling the binary):

```kLang
func main: () -> () {
    print("Hello kOS!");
}
```
which is this in kRISC instructions:

```kRISC
push ARG_MARKER
push "Hello kOS!"
call "", "print"
```
and the test kOS CPU output this:
```kOS Output
Hello kOS!
```